+++ internal/env/env_test.go
@@ -13,6 +13,7 @@ func TestConstants(t *testing.T) {
 	assert.Equal(t, "TEMPLATE_NAME", string(TemplateName))
 	assert.Equal(t, "CONFIG_FILE", string(ConfigFilePath))
 	assert.Equal(t, "DEFAULT_ERROR_PAGE", string(DefaultErrorPage))
+	assert.Equal(t, "DEFAULT_HTTP_CODE", string(DefaultHTTPCode))
 }
 
 func TestEnvVariable_Lookup(t *testing.T) {
@@ -24,6 +25,7 @@ func TestEnvVariable_Lookup(t *testing.T) {
 		{giveEnv: TemplateName},
 		{giveEnv: ConfigFilePath},
 		{giveEnv: DefaultErrorPage},
+		{giveEnv: DefaultHTTPCode},
 	}
 
 	for _, tt := range cases {